[fix] Seperate RunLogsTab and LogRow components
