+++ aim/web/ui/src/pages/RunDetail/RunLogsTab/RunLogsTab.d.ts
@@ -5,3 +5,9 @@ export interface IRunLogsTabProps {
   inProgress: boolean;
   updatedLogsCount: number;
 }
+
+export enum LogsLastRequestEnum {
+  DEFAULT = 'default',
+  LIVE_UPDATE = 'live-update',
+  LOAD_MORE = 'load-more',
+}
