+++ aim/web/ui/src/styles/abstracts/_variables.scss
@@ -22,7 +22,8 @@ $cuddle: #bdcee8;
 $denim: $primary-color;
 $black: #000000;
 $transparent: transparent;
-$grey: #333333;
+
+$grey: #586069;
 
 $grey-darker: #24292e;
 $grey-bg: #f7f7f7;
